enh: update tests
