+++ nitransforms/tests/test_io_itk.py
@@ -1,0 +1,41 @@
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+"""Test io module for ITK transforms."""
+
+import pytest
+
+import numpy as np
+import nibabel as nb
+
+from nitransforms.base import TransformError
+from nitransforms.io.base import TransformFileError
+from nitransforms.io.itk import ITKDisplacementsField
+from nitransforms.nonlinear import (
+    DenseFieldTransform,
+)
+
+
+@pytest.mark.parametrize("size", [(20, 20, 20), (20, 20, 20, 3)])
+def test_itk_disp_load(size):
+    """Checks field sizes."""
+    with pytest.raises(TransformFileError):
+        ITKDisplacementsField.from_image(
+            nb.Nifti1Image(np.zeros(size), np.eye(4), None)
+        )
+
+
+@pytest.mark.parametrize("size", [(20, 20, 20), (20, 20, 20, 2, 3), (20, 20, 20, 1, 4)])
+def test_displacements_bad_sizes(size):
+    """Checks field sizes."""
+    with pytest.raises(TransformError):
+        DenseFieldTransform(nb.Nifti1Image(np.zeros(size), np.eye(4), None))
+
+
+def test_itk_disp_load_intent():
+    """Checks whether the NIfTI intent is fixed."""
+    with pytest.warns(UserWarning):
+        field = ITKDisplacementsField.from_image(
+            nb.Nifti1Image(np.zeros((20, 20, 20, 1, 3)), np.eye(4), None)
+        )
+
+    assert field.header.get_intent()[0] == "vector"
